refactor(storage): split chunk client into object-store and chunk-client packages
pkg/storage/chunk/client conflated two things: a generic object-store
abstraction (ObjectClient, ObjectAttributes, StorageObject,
StorageCommonPrefix, PrefixedObjectClient - no dependency on
pkg/storage/chunk) and a chunk-storage wrapper on top of it (Client,
NewClient/NewClientWithMaxParallel, KeyEncoder/FSEncoder,
MetricsChunkClient - depends on pkg/storage/chunk and
pkg/storage/config).

Splits them: the chunk-aware half moves to a new package,
pkg/storage/chunk/chunkclient; pkg/storage/chunk/client keeps its
import path but now only has the generic half.

This is step 1 of breaking pkg/storage/bucket's dependency on
pkg/storage/chunk. bucket only ever needed the generic object-store
types, but importing chunk/client as one package dragged in the whole
chunk graph (chunk, chunk/cache, chunk/cache/resultscache,
storage/config) along with it. Confirmed via go list -deps that bucket
has dropped its transitive dependency on storage/config; it still
depends on pkg/storage/chunk via chunk/client/util (imported by the
aws/gcp backends), which is the next package to split the same way.

The move is code-verbatim aside from renaming the unexported client
struct to chunkClient, since the new package now imports the sibling
client package by that name.
