+++ pkg/compactor/compactor.go
@@ -24,6 +24,7 @@ import (
 	"github.com/grafana/loki/v3/pkg/compactor/jobqueue"
 	"github.com/grafana/loki/v3/pkg/compactor/retention"
 	"github.com/grafana/loki/v3/pkg/storage/bucket"
+	"github.com/grafana/loki/v3/pkg/storage/chunk/chunkclient"
 	"github.com/grafana/loki/v3/pkg/storage/chunk/client"
 	"github.com/grafana/loki/v3/pkg/storage/chunk/client/local"
 	chunk_util "github.com/grafana/loki/v3/pkg/storage/chunk/client/util"
@@ -236,7 +237,7 @@ func (c *Compactor) init(
 		if c.cfg.RetentionEnabled {
 			var (
 				raw              client.ObjectClient
-				encoder          client.KeyEncoder
+				encoder          chunkclient.KeyEncoder
 				name             = fmt.Sprintf("%s_%s", period.ObjectType, period.From.String())
 				retentionWorkDir = filepath.Join(c.cfg.WorkingDirectory, "retention", name, MarkersFolder)
 				r                = prometheus.WrapRegistererWith(prometheus.Labels{"from": name}, r)
@@ -273,11 +274,11 @@ func (c *Compactor) init(
 			// object client and to the thanos object store adapter backed by a
 			// filesystem bucket.
 			if _, ok := raw.(*local.FSObjectClient); ok {
-				encoder = client.FSEncoder
+				encoder = chunkclient.FSEncoder
 			} else if adapter, ok := raw.(*bucket.ObjectClientAdapter); ok && adapter.IsBackendFilesystem() {
-				encoder = client.FSEncoder
+				encoder = chunkclient.FSEncoder
 			}
-			chunkClient := client.NewClient(objectClient, encoder, schemaConfig)
+			chunkClient := chunkclient.NewClient(objectClient, encoder, schemaConfig)
 
 			if c.cfg.DeletionMarkerObjectStorePrefix != "" {
 				markerStorageClient = client.NewPrefixedObjectClient(raw, c.cfg.DeletionMarkerObjectStorePrefix)
