+++ pkg/compactor/deletion/job_runner.go
@@ -20,13 +20,13 @@ import (
 	"github.com/grafana/loki/v3/pkg/compactor/deletion/deletionproto"
 	"github.com/grafana/loki/v3/pkg/compactor/retention"
 	storage_chunk "github.com/grafana/loki/v3/pkg/storage/chunk"
-	"github.com/grafana/loki/v3/pkg/storage/chunk/client"
+	"github.com/grafana/loki/v3/pkg/storage/chunk/chunkclient"
 	"github.com/grafana/loki/v3/pkg/util"
 	"github.com/grafana/loki/v3/pkg/util/filter"
 	util_log "github.com/grafana/loki/v3/pkg/util/log"
 )
 
-type GetChunkClientForTableFunc func(table string) (client.Client, error)
+type GetChunkClientForTableFunc func(table string) (chunkclient.Client, error)
 
 type Chunk interface {
 	GetFrom() model.Time
