+++ pkg/storage/chunk/chunkclient/metrics.go
@@ -1,4 +1,4 @@
-package client
+package chunkclient
 
 import (
 	"context"
