+++ pkg/storage/chunk/chunkclient/object_client_test.go
@@ -1,4 +1,4 @@
-package client
+package chunkclient
 
 import (
 	"testing"
